// src/app/modules/product/product.interface.ts
export type TDesktop = {
  desktopId?: string;
  departmentName: string;
  labName: string;
  isActive: boolean;
};
export type TWhiteboard = {
  whiteboardId?: string;
  departmentName: string;
  isActive: boolean;
};
export type TMarker = {
  markerId?: string;
  departmentName: string;
  isActive: boolean;
};
export type TChair = {
  chairId?: string;
  departmentName: string;
  isActive: boolean;
};
export type TDeskBox = {
  deskBoxId?: string;
  departmentName: string;
  isActive: boolean;
};
export type TTable = {
  tableId?: string;
  departmentName: string;
  isActive: boolean;
};